refactor: 유틸 분리 및 타입 변경
