+++ src/utils/ts/date.ts
@@ -23,3 +23,10 @@ export const formatIsoDateToYYYYMMDD = (value: string): string => {
 
   return `${year}.${month.padStart(2, '0')}.${day.padStart(2, '0')}`;
 };
+
+export function formatDateDot(date: Date): string {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}.${month}.${day}`;
+}
